requirements.txt
### Library dependencies for the python code.  You need to install these with ###
### `pip install -r requirements.txt` before you can run this code            ###
#################################################################################

#### ESSENTIAL LIBRARIES FOR MAIN FUNCTIONALITY ####

# YAML parser and emitter for the Python
pyyaml
python-yaml
